rag_chatbot/requirements.txt
# Weaviate
weaviate-client==4.9.3

# AI/ML
openai==1.59.5

# Utilities
python-dotenv==1.0.0